requirements.txt
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
qdrant-client==1.9.2
pillow==10.3.0
requests==2.32.3